sentry - move main's `fn run` to `mod application`
